requirements.txt
Brotli==1.1.0
cffi==1.17.1
cryptography==44.0.1
cssselect2==0.7.0
docopt==0.6.2
fonttools==4.56.0
Jinja2==3.1.6
MarkupSafe==3.0.2
num2words==0.5.14
pdf2image==1.17.0
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.0.1
qrcode==8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
